+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Bug fixes
 - PIM-3308: Fix regression on unclassified filter
-- PIM-3311: fix creation of products with missing identifier during imports
+- PIM-3311: Fix creation of products with missing identifier during imports
+- PIM-3312: Fix CSV import of product values with invalid channel, locale or currency
 
 # 1.2.10 (2014-10-24)